+++ src/main/java/com/tristankechlo/livingthings/events/BlockEventHandler.java
@@ -2,12 +2,17 @@ package com.tristankechlo.livingthings.events;
 
 import com.tristankechlo.livingthings.config.LivingThingsConfig;
 import com.tristankechlo.livingthings.entities.AncientBlazeEntity;
+import com.tristankechlo.livingthings.entities.BabyEnderDragonEntity;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.init.ModItems;
 import com.tristankechlo.livingthings.misc.LivingThingsTags;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction.Axis;
+import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LightningBolt;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -17,6 +22,7 @@ import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.RotatedPillarBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -58,11 +64,7 @@ public class BlockEventHandler {
 	}
 
 	@SubscribeEvent
-	/** used to spawn the ancient blaze */
 	public void onBlockPlaceEvent(final BlockEvent.EntityPlaceEvent event) {
-		if (!LivingThingsConfig.ANCIENT_BLAZE.canSpawn.get()) {
-			return;
-		}
 		final LevelAccessor world = event.getWorld();
 		if (world == null) {
 			return;
@@ -77,19 +79,59 @@ public class BlockEventHandler {
 		}
 		final BlockPos pos = event.getPos();
 		final BlockState placedBlock = event.getPlacedBlock();
+		final BlockState blockBelow = world.getBlockState(pos.below());
+
 		if (placedBlock.is(Blocks.JACK_O_LANTERN) && world.getBlockState(pos.below()).is(Blocks.GLOWSTONE)
 				&& world.getBlockState(pos.below(2)).is(Blocks.GLOWSTONE)) {
 
+			if (!LivingThingsConfig.ANCIENT_BLAZE.canSpawn.get()) {
+				return;
+			}
+
+			// remove blocks
 			world.setBlock(pos, Blocks.AIR.defaultBlockState(), 3);
 			world.setBlock(pos.below(), Blocks.AIR.defaultBlockState(), 3);
 			world.setBlock(pos.below(2), Blocks.AIR.defaultBlockState(), 3);
 
+			// spawn ancient blaze
 			AncientBlazeEntity blaze = ModEntityTypes.ANCIENT_BLAZE.get().create((Level) world);
 			blaze.setInvulnerableTime(LivingThingsConfig.ANCIENT_BLAZE.chargingTime.get());
 			blaze.setPos(pos.getX() + 0.5D, pos.below(2).getY() + 0.2D, pos.getZ() + 0.5D);
 			world.addFreshEntity(blaze);
 			return;
 
+		} else if (placedBlock.is(Blocks.DRAGON_EGG) && blockBelow.is(Blocks.PURPUR_PILLAR)
+				&& blockBelow.getValue(RotatedPillarBlock.AXIS) == Axis.Y) {
+
+			if (world.dimensionType().hasCeiling()) {
+				return;
+			}
+
+			if (!LivingThingsConfig.BABY_ENDER_DRAGON.canSpawn.get()) {
+				return;
+			}
+
+			// remove dragon egg
+			world.setBlock(pos, Blocks.AIR.defaultBlockState(), 3);
+
+			// summon some lightning bolts
+			final int count = 10;
+			final double radius = 6.0D;
+			for (int i = 0; i < count; i++) {
+				LightningBolt bolt = EntityType.LIGHTNING_BOLT.create((Level) world);
+				final double x = pos.getX() + (radius * Mth.cos((i * 2 * Mth.PI) / count));
+				final double z = pos.getZ() + (radius * Mth.sin((i * 2 * Mth.PI) / count));
+				final double y = pos.below().getY();
+				bolt.setPos(x, y, z);
+				world.addFreshEntity(bolt);
+			}
+
+			// spawn baby ender dragon
+			BabyEnderDragonEntity dragon = ModEntityTypes.BABY_ENDER_DRAGON.get().create((Level) world);
+			dragon.setPos(pos.getX() + 0.5D, pos.getY() + 0.1D, pos.getZ() + 0.5D);
+			world.addFreshEntity(dragon);
+			return;
 		}
 	}
+
 }